+++ packages/rsnap-overlay/src/overlay.rs
@@ -7,6 +7,8 @@ mod config_runtime;
 mod cursor_context_runtime;
 mod cursor_runtime;
 mod frozen_brush_runtime;
+mod frozen_export_runtime;
+mod frozen_mosaic_runtime;
 mod frozen_selection_runtime;
 mod hud_helpers;
 mod hud_runtime;
@@ -64,10 +66,9 @@ use egui::{
 };
 use egui_phosphor::{Variant, regular};
 use egui_wgpu::{Renderer, ScreenDescriptor};
-use image::{
-	RgbaImage,
-	imageops::{self, FilterType},
-};
+use image::RgbaImage;
+#[cfg(target_os = "macos")]
+use image::imageops;
 #[cfg(target_os = "macos")]
 use objc::declare::ClassDecl;
 #[cfg(target_os = "macos")]
@@ -181,7 +182,6 @@ use crate::deferred_text_recognition::DeferredTextRecognitionRequest;
 use crate::live_frame_stream_macos::{CursorSampleRequest, MacLiveFrameStream};
 use crate::scroll_capture::{self, ScrollDirection, ScrollObserveOutcome, ScrollSession};
 use crate::state::LiveCursorSample;
-use crate::text_rendering::{self, RasterTextAnnotation};
 use crate::worker::CapturedMonitorRegionResult;
 use crate::{
 	state::{
@@ -1335,358 +1335,6 @@ impl OverlaySession {
 		}
 	}
 
-	fn cropped_frozen_capture_image(&self) -> Option<RgbaImage> {
-		if self.frozen_capture_source != FrozenCaptureSource::FullscreenFallback
-			&& let Some(window_image) = self.frozen_window_image.as_ref()
-		{
-			match self.config.window_capture_alpha_mode {
-				WindowCaptureAlphaMode::Background => {},
-				WindowCaptureAlphaMode::MatteLight => {
-					return Some(window_image.clone());
-				},
-				WindowCaptureAlphaMode::MatteDark => {
-					return Some(window_image.clone());
-				},
-			}
-		}
-
-		let frozen_image = self.state.frozen_image.as_ref()?;
-		let Some(monitor) = self.state.monitor else {
-			return Some(frozen_image.clone());
-		};
-		let capture_rect = self
-			.state
-			.frozen_capture_rect
-			.unwrap_or_else(|| RectPoints::new(0, 0, monitor.width, monitor.height));
-		let capture_rect = monitor.local_rect_to_pixels(capture_rect);
-		let x = capture_rect.x.min(frozen_image.width());
-		let y = capture_rect.y.min(frozen_image.height());
-		let max_width = frozen_image.width().saturating_sub(x);
-		let max_height = frozen_image.height().saturating_sub(y);
-		let width = capture_rect.width.min(max_width);
-		let height = capture_rect.height.min(max_height);
-
-		if width == 0 || height == 0 {
-			None
-		} else {
-			Some(imageops::crop_imm(frozen_image, x, y, width, height).to_image())
-		}
-	}
-
-	fn render_frozen_committed_overlays_into_image(&self, image: &mut RgbaImage) {
-		if self.scroll_capture.active {
-			return;
-		}
-
-		let Some(export_transform) = self.frozen_export_transform_for_image(image) else {
-			return;
-		};
-		let mut brush_coverage_mask = None;
-
-		Self::for_each_frozen_committed_overlay(
-			&self.frozen_edit_undo_stack,
-			&self.frozen_brush.committed_strokes,
-			&self.frozen_text_annotations,
-			|overlay| match overlay {
-				FrozenCommittedOverlay::Brush(stroke) => {
-					let coverage_mask = brush_coverage_mask.get_or_insert_with(|| {
-						vec![0_u8; image.width() as usize * image.height() as usize]
-					});
-
-					Self::rasterize_frozen_brush_points_into_image(
-						image,
-						coverage_mask,
-						export_transform,
-						&stroke.points,
-					);
-				},
-				FrozenCommittedOverlay::Text(annotation) => {
-					Self::render_frozen_text_annotation_into_image(
-						image,
-						export_transform,
-						annotation,
-					);
-				},
-			},
-		);
-
-		if let Some(active_stroke) = &self.frozen_brush.active_stroke {
-			let display_points = Self::active_frozen_brush_display_points(active_stroke);
-			let coverage_mask = brush_coverage_mask.get_or_insert_with(|| {
-				vec![0_u8; image.width() as usize * image.height() as usize]
-			});
-
-			Self::rasterize_frozen_brush_points_into_image(
-				image,
-				coverage_mask,
-				export_transform,
-				&display_points,
-			);
-		}
-	}
-
-	fn rasterize_frozen_brush_points_into_image(
-		export_image: &mut RgbaImage,
-		coverage_mask: &mut [u8],
-		export_transform: FrozenExportTransform,
-		points: &[Pos2],
-	) {
-		if export_image.width() == 0 || export_image.height() == 0 {
-			return;
-		}
-		if coverage_mask.len() != export_image.width() as usize * export_image.height() as usize {
-			return;
-		}
-
-		let radius =
-			(FROZEN_BRUSH_STROKE_WIDTH_POINTS * export_transform.scalar_scale() * 0.5).max(1.0);
-		let color = image::Rgba(FROZEN_BRUSH_COLOR_RGBA);
-
-		coverage_mask.fill(0);
-
-		Self::rasterize_frozen_brush_points(
-			coverage_mask,
-			export_image.width(),
-			export_image.height(),
-			points,
-			export_transform,
-			radius,
-		);
-		Self::blend_frozen_brush_coverage_mask(export_image, coverage_mask, color);
-	}
-
-	fn for_each_frozen_committed_overlay(
-		edit_history: &[FrozenEditKind],
-		brush_strokes: &[FrozenBrushStroke],
-		text_annotations: &[FrozenTextAnnotation],
-		mut f: impl FnMut(FrozenCommittedOverlay<'_>),
-	) {
-		let mut brush_index = 0;
-		let mut text_index = 0;
-
-		for edit_kind in edit_history {
-			match edit_kind {
-				FrozenEditKind::BrushStroke => {
-					let Some(stroke) = brush_strokes.get(brush_index) else {
-						continue;
-					};
-
-					brush_index += 1;
-
-					f(FrozenCommittedOverlay::Brush(stroke));
-				},
-				FrozenEditKind::TextAnnotation => {
-					let Some(annotation) = text_annotations.get(text_index) else {
-						continue;
-					};
-
-					text_index += 1;
-
-					f(FrozenCommittedOverlay::Text(annotation));
-				},
-				FrozenEditKind::MosaicEdit => {},
-			}
-		}
-	}
-
-	fn rasterize_frozen_brush_points(
-		coverage_mask: &mut [u8],
-		export_width: u32,
-		export_height: u32,
-		points: &[Pos2],
-		export_transform: FrozenExportTransform,
-		radius: f32,
-	) {
-		let rendered_points =
-			Self::rendered_frozen_brush_points(points, FROZEN_BRUSH_RENDER_SAMPLE_STEP_POINTS);
-		let Some(first) = rendered_points.first().copied() else {
-			return;
-		};
-		let mut previous = export_transform.point_to_pixels(first);
-
-		Self::rasterize_frozen_brush_circle(
-			coverage_mask,
-			export_width,
-			export_height,
-			previous,
-			radius,
-		);
-
-		for point in rendered_points.into_iter().skip(1) {
-			let current = export_transform.point_to_pixels(point);
-
-			Self::rasterize_frozen_brush_segment(
-				coverage_mask,
-				export_width,
-				export_height,
-				previous,
-				current,
-				radius,
-			);
-
-			previous = current;
-		}
-	}
-
-	fn frozen_export_capture_rect(&self) -> Option<RectPoints> {
-		self.state.frozen_capture_rect.or_else(|| {
-			self.state.monitor.map(|monitor| RectPoints::new(0, 0, monitor.width, monitor.height))
-		})
-	}
-
-	fn frozen_export_transform_for_image(
-		&self,
-		image: &RgbaImage,
-	) -> Option<FrozenExportTransform> {
-		FrozenExportTransform::new(
-			self.frozen_export_capture_rect()?,
-			image.width(),
-			image.height(),
-		)
-	}
-
-	fn rasterize_frozen_brush_segment(
-		coverage_mask: &mut [u8],
-		export_width: u32,
-		export_height: u32,
-		start: Pos2,
-		end: Pos2,
-		radius: f32,
-	) {
-		let delta = end - start;
-		let delta_len_sq = delta.length_sq();
-
-		if delta_len_sq <= f32::EPSILON {
-			Self::rasterize_frozen_brush_circle(
-				coverage_mask,
-				export_width,
-				export_height,
-				start,
-				radius,
-			);
-
-			return;
-		}
-
-		let min_x = ((start.x.min(end.x) - radius - 0.5).floor().max(0.0)) as u32;
-		let min_y = ((start.y.min(end.y) - radius - 0.5).floor().max(0.0)) as u32;
-		let max_x = ((start.x.max(end.x) + radius + 0.5)
-			.ceil()
-			.min((export_width.saturating_sub(1)) as f32)) as u32;
-		let max_y = ((start.y.max(end.y) + radius + 0.5)
-			.ceil()
-			.min((export_height.saturating_sub(1)) as f32)) as u32;
-
-		for y in min_y..=max_y {
-			for x in min_x..=max_x {
-				let sample = Pos2::new(x as f32 + 0.5, y as f32 + 0.5);
-				let projection = ((sample - start).dot(delta) / delta_len_sq).clamp(0.0, 1.0);
-				let nearest = start + delta * projection;
-				let coverage = Self::frozen_brush_coverage(sample.distance(nearest), radius);
-
-				Self::update_frozen_brush_coverage_mask(
-					coverage_mask,
-					export_width,
-					x,
-					y,
-					coverage,
-				);
-			}
-		}
-	}
-
-	fn rasterize_frozen_brush_circle(
-		coverage_mask: &mut [u8],
-		export_width: u32,
-		export_height: u32,
-		center: Pos2,
-		radius: f32,
-	) {
-		if export_width == 0 || export_height == 0 {
-			return;
-		}
-
-		let min_x = ((center.x - radius - 0.5).floor().max(0.0)) as u32;
-		let min_y = ((center.y - radius - 0.5).floor().max(0.0)) as u32;
-		let max_x =
-			(center.x + radius + 0.5).ceil().min((export_width.saturating_sub(1)) as f32) as u32;
-		let max_y =
-			(center.y + radius + 0.5).ceil().min((export_height.saturating_sub(1)) as f32) as u32;
-
-		if min_x > max_x || min_y > max_y {
-			return;
-		}
-
-		for y in min_y..=max_y {
-			for x in min_x..=max_x {
-				let sample = Pos2::new(x as f32 + 0.5, y as f32 + 0.5);
-				let coverage = Self::frozen_brush_coverage(sample.distance(center), radius);
-
-				Self::update_frozen_brush_coverage_mask(
-					coverage_mask,
-					export_width,
-					x,
-					y,
-					coverage,
-				);
-			}
-		}
-	}
-
-	fn frozen_brush_coverage(distance: f32, radius: f32) -> u8 {
-		((radius + 0.5 - distance).clamp(0.0, 1.0) * 255.0).round() as u8
-	}
-
-	fn update_frozen_brush_coverage_mask(
-		coverage_mask: &mut [u8],
-		export_width: u32,
-		x: u32,
-		y: u32,
-		coverage: u8,
-	) {
-		if coverage == 0 {
-			return;
-		}
-
-		let index = y as usize * export_width as usize + x as usize;
-
-		coverage_mask[index] = coverage_mask[index].max(coverage);
-	}
-
-	fn blend_frozen_brush_coverage_mask(
-		export_image: &mut RgbaImage,
-		coverage_mask: &[u8],
-		color: image::Rgba<u8>,
-	) {
-		let source_alpha = color[3] as f32 / 255.0;
-
-		for (index, pixel) in export_image.pixels_mut().enumerate() {
-			let mask_alpha = coverage_mask[index];
-
-			if mask_alpha == 0 {
-				continue;
-			}
-
-			let src_a = (mask_alpha as f32 / 255.0) * source_alpha;
-			let dst_a = pixel[3] as f32 / 255.0;
-			let out_a = src_a + dst_a * (1.0 - src_a);
-
-			if out_a <= f32::EPSILON {
-				continue;
-			}
-
-			for channel in 0..3 {
-				let src = color[channel] as f32 / 255.0;
-				let dst = pixel[channel] as f32 / 255.0;
-				let out = (src * src_a + dst * dst_a * (1.0 - src_a)) / out_a;
-
-				pixel[channel] = (out * 255.0).round().clamp(0.0, 255.0) as u8;
-			}
-
-			pixel[3] = (out_a * 255.0).round().clamp(0.0, 255.0) as u8;
-		}
-	}
-
 	#[cfg(target_os = "macos")]
 	fn cropped_monitor_frozen_region_image(
 		&self,
@@ -1713,126 +1361,6 @@ impl OverlaySession {
 		} else {
 			Some(imageops::crop_imm(frozen_image, x, y, width, height).to_image())
 		}
-	}
-
-	fn intersect_rect_points(left: RectPoints, right: RectPoints) -> Option<RectPoints> {
-		let x = left.x.max(right.x);
-		let y = left.y.max(right.y);
-		let right_edge = left.x.saturating_add(left.width).min(right.x.saturating_add(right.width));
-		let bottom_edge =
-			left.y.saturating_add(left.height).min(right.y.saturating_add(right.height));
-
-		(right_edge > x && bottom_edge > y).then(|| {
-			RectPoints::new(x, y, right_edge.saturating_sub(x), bottom_edge.saturating_sub(y))
-		})
-	}
-
-	fn build_frozen_image_patch(image: &RgbaImage, rect: RectPoints) -> Option<FrozenImagePatch> {
-		let x = rect.x.min(image.width());
-		let y = rect.y.min(image.height());
-		let max_width = image.width().saturating_sub(x);
-		let max_height = image.height().saturating_sub(y);
-		let width = rect.width.min(max_width);
-		let height = rect.height.min(max_height);
-
-		if width == 0 || height == 0 {
-			return None;
-		}
-
-		let rect = RectPoints::new(x, y, width, height);
-		let before = imageops::crop_imm(image, x, y, width, height).to_image();
-		let after = Self::mosaic_patch(&before);
-
-		Some(FrozenImagePatch { rect, before, after })
-	}
-
-	fn mosaic_patch(region: &RgbaImage) -> RgbaImage {
-		let mut out = region.clone();
-
-		for block_y in (0..region.height()).step_by(FROZEN_MOSAIC_BLOCK_SIZE_PX as usize) {
-			for block_x in (0..region.width()).step_by(FROZEN_MOSAIC_BLOCK_SIZE_PX as usize) {
-				let block_width = FROZEN_MOSAIC_BLOCK_SIZE_PX.min(region.width() - block_x);
-				let block_height = FROZEN_MOSAIC_BLOCK_SIZE_PX.min(region.height() - block_y);
-				let mut sum = [0_u64; 4];
-				let mut samples = 0_u64;
-
-				for y in block_y..block_y.saturating_add(block_height) {
-					for x in block_x..block_x.saturating_add(block_width) {
-						let pixel = region.get_pixel(x, y);
-
-						sum[0] = sum[0].saturating_add(u64::from(pixel[0]));
-						sum[1] = sum[1].saturating_add(u64::from(pixel[1]));
-						sum[2] = sum[2].saturating_add(u64::from(pixel[2]));
-						sum[3] = sum[3].saturating_add(u64::from(pixel[3]));
-						samples = samples.saturating_add(1);
-					}
-				}
-
-				if samples == 0 {
-					continue;
-				}
-
-				let fill = image::Rgba([
-					(sum[0] / samples) as u8,
-					(sum[1] / samples) as u8,
-					(sum[2] / samples) as u8,
-					(sum[3] / samples) as u8,
-				]);
-
-				for y in block_y..block_y.saturating_add(block_height) {
-					for x in block_x..block_x.saturating_add(block_width) {
-						out.put_pixel(x, y, fill);
-					}
-				}
-			}
-		}
-
-		out
-	}
-
-	fn apply_frozen_image_patch(image: &mut RgbaImage, patch: &FrozenImagePatch, use_after: bool) {
-		let source = if use_after { &patch.after } else { &patch.before };
-
-		imageops::replace(image, source, i64::from(patch.rect.x), i64::from(patch.rect.y));
-	}
-
-	fn map_rect_into_window_image(
-		monitor: MonitorRect,
-		capture_rect_points: RectPoints,
-		window_image: &RgbaImage,
-		selection_rect_points: RectPoints,
-	) -> Option<RectPoints> {
-		let capture_rect_px = monitor.local_rect_to_pixels(capture_rect_points);
-		let selection_rect_px = monitor.local_rect_to_pixels(selection_rect_points);
-		let selection_rect_px = Self::intersect_rect_points(selection_rect_px, capture_rect_px)?;
-
-		if capture_rect_px.is_empty() || window_image.width() == 0 || window_image.height() == 0 {
-			return None;
-		}
-
-		let rel_left = selection_rect_px.x.saturating_sub(capture_rect_px.x);
-		let rel_top = selection_rect_px.y.saturating_sub(capture_rect_px.y);
-		let rel_right = rel_left.saturating_add(selection_rect_px.width);
-		let rel_bottom = rel_top.saturating_add(selection_rect_px.height);
-		let capture_width = u64::from(capture_rect_px.width.max(1));
-		let capture_height = u64::from(capture_rect_px.height.max(1));
-		let target_width = u64::from(window_image.width());
-		let target_height = u64::from(window_image.height());
-		let left = (u64::from(rel_left) * target_width) / capture_width;
-		let top = (u64::from(rel_top) * target_height) / capture_height;
-		let right = (u64::from(rel_right) * target_width).div_ceil(capture_width);
-		let bottom = (u64::from(rel_bottom) * target_height).div_ceil(capture_height);
-		let width = right.saturating_sub(left) as u32;
-		let height = bottom.saturating_sub(top) as u32;
-
-		(width > 0 && height > 0).then(|| {
-			RectPoints::new(
-				left.min(target_width) as u32,
-				top.min(target_height) as u32,
-				width,
-				height,
-			)
-		})
 	}
 
 	fn note_frozen_image_mutated(&mut self, monitor: MonitorRect) {
@@ -1880,123 +1408,6 @@ impl OverlaySession {
 		}
 
 		self.clear_frozen_redo_history();
-	}
-
-	fn push_frozen_mosaic_edit(&mut self, edit: FrozenMosaicEdit) {
-		self.frozen_mosaic_undo_stack.push(edit);
-	}
-
-	fn apply_frozen_mosaic_edit(&mut self, rect_points: RectPoints) -> bool {
-		let Some(monitor) = self.state.monitor else {
-			return false;
-		};
-
-		if !self.frozen_final_capture_ready() {
-			return false;
-		}
-
-		let preview_rect_px = monitor.local_rect_to_pixels(rect_points);
-		let Some(preview_patch) = self
-			.state
-			.frozen_image
-			.as_ref()
-			.and_then(|image| Self::build_frozen_image_patch(image, preview_rect_px))
-		else {
-			return false;
-		};
-		let window_patch = match (self.frozen_window_image.as_ref(), self.state.frozen_capture_rect)
-		{
-			(Some(window_image), Some(capture_rect_points)) => Self::map_rect_into_window_image(
-				monitor,
-				capture_rect_points,
-				window_image,
-				rect_points,
-			)
-			.and_then(|rect| Self::build_frozen_image_patch(window_image, rect)),
-			_ => None,
-		};
-
-		if let Some(image) = self.state.frozen_image.as_mut() {
-			Self::apply_frozen_image_patch(image, &preview_patch, true);
-		}
-		if let (Some(window_image), Some(window_patch)) =
-			(self.frozen_window_image.as_mut(), window_patch.as_ref())
-		{
-			Self::apply_frozen_image_patch(window_image, window_patch, true);
-		}
-
-		self.push_frozen_mosaic_edit(FrozenMosaicEdit { preview_patch, window_patch });
-		self.push_frozen_edit_to_undo_history(FrozenEditKind::MosaicEdit);
-		self.note_frozen_image_mutated(monitor);
-
-		true
-	}
-
-	fn replay_frozen_mosaic_edit(&mut self, edit: &FrozenMosaicEdit, use_after: bool) -> bool {
-		let Some(monitor) = self.state.monitor else {
-			return false;
-		};
-
-		if let Some(image) = self.state.frozen_image.as_mut() {
-			Self::apply_frozen_image_patch(image, &edit.preview_patch, use_after);
-		}
-		if let (Some(window_image), Some(window_patch)) =
-			(self.frozen_window_image.as_mut(), edit.window_patch.as_ref())
-		{
-			Self::apply_frozen_image_patch(window_image, window_patch, use_after);
-		}
-
-		self.note_frozen_image_mutated(monitor);
-
-		true
-	}
-
-	fn undo_frozen_mosaic_edit(&mut self) -> bool {
-		if !self.frozen_final_capture_ready() {
-			return false;
-		}
-
-		let Some(edit) = self.frozen_mosaic_undo_stack.pop() else {
-			return false;
-		};
-		let reapplied = self.replay_frozen_mosaic_edit(&edit, false);
-
-		self.frozen_mosaic_redo_stack.push(edit);
-		self.sync_frozen_toolbar_state();
-
-		reapplied
-	}
-
-	fn redo_frozen_mosaic_edit(&mut self) -> bool {
-		if !self.frozen_final_capture_ready() {
-			return false;
-		}
-
-		let Some(edit) = self.frozen_mosaic_redo_stack.pop() else {
-			return false;
-		};
-		let reapplied = self.replay_frozen_mosaic_edit(&edit, true);
-
-		self.frozen_mosaic_undo_stack.push(edit);
-		self.sync_frozen_toolbar_state();
-
-		reapplied
-	}
-
-	fn commit_frozen_mosaic_drag(&mut self) -> bool {
-		let preview_rect = self.state.frozen_mosaic_preview_rect;
-
-		self.stop_frozen_mosaic_drag();
-
-		let Some(preview_rect) = preview_rect else {
-			return false;
-		};
-
-		if preview_rect.width <= 1 && preview_rect.height <= 1 {
-			return false;
-		}
-
-		self.apply_frozen_mosaic_edit(preview_rect)
 	}
 
 	fn frozen_undo_available(&self) -> bool {
@@ -2047,80 +1458,6 @@ impl OverlaySession {
 		self.sync_frozen_toolbar_state();
 
 		redone
-	}
-
-	fn flatten_window_image_with_matte(image: &RgbaImage, matte: image::Rgba<u8>) -> RgbaImage {
-		let mut out = image.clone();
-
-		for pixel in out.pixels_mut() {
-			let alpha = u16::from(pixel[3]);
-			let inv_alpha = 255_u16.saturating_sub(alpha);
-
-			for channel in 0..3 {
-				let src = u16::from(pixel[channel]);
-				let bg = u16::from(matte[channel]);
-				let blended =
-					(src.saturating_mul(alpha) + bg.saturating_mul(inv_alpha) + 127) / 255;
-
-				pixel[channel] = blended as u8;
-			}
-
-			pixel[3] = 255;
-		}
-
-		out
-	}
-
-	fn compose_window_preview_layer(
-		window_image: &RgbaImage,
-		alpha_mode: WindowCaptureAlphaMode,
-	) -> RgbaImage {
-		match alpha_mode {
-			WindowCaptureAlphaMode::Background => window_image.clone(),
-			WindowCaptureAlphaMode::MatteLight => {
-				Self::flatten_window_image_with_matte(window_image, WINDOW_CAPTURE_MATTE_LIGHT_RGBA)
-			},
-			WindowCaptureAlphaMode::MatteDark => {
-				Self::flatten_window_image_with_matte(window_image, WINDOW_CAPTURE_MATTE_DARK_RGBA)
-			},
-		}
-	}
-
-	fn composite_window_capture_preview(
-		mut monitor_image: RgbaImage,
-		window_image: &RgbaImage,
-		monitor: MonitorRect,
-		capture_rect_points: RectPoints,
-		alpha_mode: WindowCaptureAlphaMode,
-	) -> RgbaImage {
-		let capture_rect_px = monitor.local_rect_to_pixels(capture_rect_points);
-
-		if capture_rect_px.width == 0 || capture_rect_px.height == 0 {
-			return monitor_image;
-		}
-
-		let window_overlay = if window_image.width() == capture_rect_px.width
-			&& window_image.height() == capture_rect_px.height
-		{
-			window_image.clone()
-		} else {
-			imageops::resize(
-				window_image,
-				capture_rect_px.width,
-				capture_rect_px.height,
-				FilterType::Triangle,
-			)
-		};
-		let preview_layer = Self::compose_window_preview_layer(&window_overlay, alpha_mode);
-
-		imageops::overlay(
-			&mut monitor_image,
-			&preview_layer,
-			i64::from(capture_rect_px.x),
-			i64::from(capture_rect_px.y),
-		);
-
-		monitor_image
 	}
 
 	fn frozen_text_tool_active(&self) -> bool {
@@ -2489,21 +1826,6 @@ impl OverlaySession {
 		}
 
 		self.finish_frozen_text_editing(true)
-	}
-
-	fn render_frozen_text_annotation_into_image(
-		image: &mut RgbaImage,
-		export_transform: FrozenExportTransform,
-		annotation: &FrozenTextAnnotation,
-	) {
-		let raster_annotation = RasterTextAnnotation {
-			anchor_px: export_transform.point_to_pixels(annotation.anchor),
-			font_size_px: annotation.style.font_size_points * export_transform.scalar_scale(),
-			fill_rgba: annotation.style.color.export_rgba(),
-			text: annotation.text.as_str(),
-		};
-
-		text_rendering::render_text_annotations(image, &[raster_annotation]);
 	}
 
 	fn handle_captured_freeze_response(
@@ -2910,23 +2232,6 @@ impl OverlaySession {
 			left_button_went_down,
 			left_button_went_up,
 		})
-	}
-
-	fn current_export_image(&self) -> Option<RgbaImage> {
-		if self.scroll_capture.active {
-			return self
-				.scroll_capture
-				.session
-				.as_ref()
-				.map(|session| session.export_image().clone());
-		}
-
-		let mut export_image =
-			self.cropped_frozen_capture_image().or_else(|| self.state.frozen_image.clone())?;
-
-		self.render_frozen_committed_overlays_into_image(&mut export_image);
-
-		Some(export_image)
 	}
 
 	#[cfg(test)]
